Removes expectation of a precise message
The tests still expect the `new` method to be called, and to return a value.

It simply no longer strictly expects the `new` method to be called with a specific paragraph of text.

This allows for flexibility in changing the text of the message, while still testing that the methods are being called.